update cron job with new Logic
