+++ src/lib/db.tsx
@@ -2,4 +2,5 @@ export enum Tables {
   Members = "members",
   Attendance = "attendance",
   Leaves = "leaves",
+  Notification = 'notifications'
 }
